fix: refine not found handling

+++ src/controllers/banner.controller.ts
@@ -2,7 +2,7 @@ import { Response } from "express";
 import { IPaginationQuery, IReqUser } from "../utils/interfaces";
 import response from "../utils/response";
 import BannerModel, { bannerDAO, TypeBanner } from "../models/banner.model";
-import { FilterQuery } from "mongoose";
+import { FilterQuery, isValidObjectId } from "mongoose";
 
 export default {
   async create(req: IReqUser, res: Response) {
@@ -58,20 +58,30 @@ export default {
   async findOne(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+
+      if (!isValidObjectId(id)) {
+        return response.notfound(res, "Failed to find a banner");
+      }
+
       const result = await BannerModel.findById(id);
 
       if (!result) {
-        return response.notfound(res, "Failed to find one a banner");
+        return response.notfound(res, "Failed to find a banner");
       }
 
-      response.success(res, result, "Success find one a banner");
+      response.success(res, result, "Success find a banner");
     } catch (error) {
-      response.error(res, error, "Failed to find one a banner");
+      response.error(res, error, "Failed to find a banner");
     }
   },
   async update(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+
+      if (!isValidObjectId(id)) {
+        return response.notfound(res, "Failed to update a banner");
+      }
+
       const result = await BannerModel.findByIdAndUpdate(id, req.body, {
         new: true,
       });
@@ -83,6 +93,11 @@ export default {
   async remove(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+
+      if (!isValidObjectId(id)) {
+        return response.notfound(res, "Failed to remove a banner");
+      }
+
       const result = await BannerModel.findByIdAndDelete(id, {
         new: true,
       });
