+++ src/controllers/ticket.controller.ts
@@ -59,24 +59,34 @@ export default {
   async findOne(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+
+      if (!isValidObjectId(id)) {
+        return response.notfound(res, "Failed to find a ticket");
+      }
+
       const result = await TicketModel.findById(id);
 
       if (!result) {
-        return response.notfound(res, "Failed to find one ticket");
+        return response.notfound(res, "Failed to find a ticket");
       }
 
-      response.success(res, result, "Success find one a ticket");
+      response.success(res, result, "Success find a ticket");
     } catch (error) {
-      response.error(res, error, "Failed to find one a ticket");
+      response.error(res, error, "Failed to find a ticket");
     }
   },
   async update(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+
+      if (!isValidObjectId(id)) {
+        return response.notfound(res, "Failed to update a ticket");
+      }
+
       const result = await TicketModel.findByIdAndUpdate(id, req.body, {
         new: true,
       });
-      response.success(res, result, "Success update one a ticket");
+      response.success(res, result, "Success update a ticket");
     } catch (error) {
       response.error(res, error, "Failed to update a ticket");
     }
@@ -84,6 +94,11 @@ export default {
   async remove(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+
+      if (!isValidObjectId(id)) {
+        return response.notfound(res, "Failed to remove a ticket");
+      }
+
       const result = await TicketModel.findByIdAndDelete(id, {
         new: true,
       });
